ci: update name-tests-test and rename all necessary files
